Moved taborder.js into src/ui/tools
