+++ src/ui/components/navigation/navigationcomponent.js
@@ -6,7 +6,7 @@ import Component from '../component';
 import { AnswersComponentError } from '../../../core/errors/errors';
 import StorageKeys from '../../../core/storage/storagekeys';
 import DOM from '../../dom/dom';
-import { mergeTabOrder, getDefaultTabOrder, getUrlParams } from '../../../core/utils/taborder';
+import { mergeTabOrder, getDefaultTabOrder, getUrlParams } from '../../tools/taborder';
 import { addParamsToUrl } from '../../../core/utils/urlutils';
 
 /**
